refactor: move schema reducer, rename types
