+++ src/containers/Tenant/Tenant.tsx
@@ -10,7 +10,7 @@ import {DEFAULT_IS_TENANT_SUMMARY_COLLAPSED, DEFAULT_SIZE_TENANT_KEY} from '../.
 import {useTypedSelector} from '../../utils/hooks';
 import routes, {CLUSTER_PAGES, createHref} from '../../routes';
 import {setHeader} from '../../store/reducers/header';
-import {disableAutorefresh, getSchema, resetLoadingState} from '../../store/reducers/schema';
+import {disableAutorefresh, getSchema, resetLoadingState} from '../../store/reducers/schema/schema';
 import {getSchemaAcl} from '../../store/reducers/schemaAcl/schemaAcl';
 import {getTenantInfo, clearTenant} from '../../store/reducers/tenant/tenant';
